+++ test_integration/struct_test.exs
@@ -23,4 +23,15 @@ defmodule Mneme.Integration.StructTest do
     # y
     auto_assert %{__struct__: Foo} <- %{__struct__: Foo}
   end
+
+  test "structs as map keys do not generate empty pattern" do
+    m = %{%User{name: "name"} => :foo}
+
+    # y
+    auto_assert ^m <- m
+    # k y
+    auto_assert %{%User{name: "name"} => :foo} <- m
+    # k k y
+    auto_assert ^m <- m
+  end
 end