+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This format is based on [Keep a Changelog](https://keepachangelog.com) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+  * Don't simplify structs used as keys in maps ([#96](https://github.com/zachallaun/mneme/issues/96)).
+
 ## v0.9.2 (2024-10-14)
 
 ### Added